requirements.txt
fastapi
uvicorn
requests
langchain
langchain-community
python-dotenv
